Renamed InputValidationResult to ValidationResult so it´s usable with other validation types
