+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/ValidationError.cs
@@ -12,19 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.ComponentModel.DataAnnotations;
-
-namespace Energinet.DataHub.MeteringPoints.Application.InputValidation
+namespace Energinet.DataHub.MeteringPoints.Application.Validation
 {
-    /// <summary>
-    /// Validator interface
-    /// </summary>
-    public interface IValidator<in TCommand, out TResponse>
+    public class ValidationError
     {
-        /// <summary>
-        /// Perform validation
-        /// </summary>
-        /// <returns>A <see cref="ValidationResult"/> representing the result of the asynchronous operation.</returns>
-        InputValidationResult Validate(TCommand command);
+        public ValidationError(string key, string message)
+        {
+            Key = key;
+            Message = message;
+        }
+
+        public string Key { get; }
+
+        public string Message { get; }
     }
 }
