+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/ContainerExtensions.cs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MeteringPoints.Application.InputValidation;
-using Energinet.DataHub.MeteringPoints.Application.InputValidation.Validators;
+using Energinet.DataHub.MeteringPoints.Application.Validation;
+using Energinet.DataHub.MeteringPoints.Application.Validation.InputValidators;
 using SimpleInjector;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure
